image-to-image-video-processorの分離
