+++ packages/image-to-image-video-processor/.prettierrc.json
@@ -1,0 +1,5 @@
+{
+  "printWidth": 120,
+  "semi": true,
+  "trailingComma": "es5"
+}
