+++ packages/image-to-image-video-processor/README.md
@@ -1,0 +1,27 @@
+# @shiguredo/image-to-image-video-processor
+
+[![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
+
+画像→画像変換を行う映像トラック処理用の内部向けユーティリティパッケージ。
+ユーザが直接利用することは想定されておらず npm にも公開されていない。
+
+## ライセンス
+
+[Apache License 2.0](https://www.apache.org/licenses/LICENSE-2.0)
+
+```
+Copyright 2023-2023, Takeru Ohta (Original Author)
+Copyright 2023-2023, Shiguredo Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+```
