+++ packages/image-to-image-video-processor/typedoc.json
@@ -1,0 +1,10 @@
+{
+  "entryPoints": ["src/video_track_processor.ts"],
+  "tsconfig": "./tsconfig.json",
+  "disableSources": true,
+  "excludePrivate": true,
+  "excludeProtected": true,
+  "excludeInternal": true,
+  "readme": "./README.md",
+  "out": "apidoc"
+}
